+++ py/desisurvey/NTS.py
@@ -441,14 +441,14 @@ class NTS():
             self.log.info('No time specified, using current time, MJD: %f' %
                           mjd)
 
-        self.queuedlist.restore()
-        if len(self.queuedlist.queued) == 0:
-            current_ra = None
-            current_dec = None
-
         previoustiles = exposure.get('previoustiles', [])
         if previoustiles is None:
             previoustiles = []
+        self.queuedlist.restore()
+        if (len(self.queuedlist.queued) == 0) and (len(previoustiles) == 0):
+            current_ra = None
+            current_dec = None
+
         previoustiles = previoustiles + self.queuedlist.queued
         # remove previous tiles from possible tiles to schedule
         ind, mask = self.scheduler.tiles.index(previoustiles,